dashboard/requirements.txt
openai
instructor
pydantic
python-dotenv
langchain
google-genai
opencv-python
numpy
supabase
uvicorn 
fastapi